move pair_analysis into ui/pair_analysis
